+++ NEWS.md
@@ -4,6 +4,7 @@
 * Fix: Added internal workaround for `PipeOpNMF` attaching `Biobase`, `BiocGenerics`, and `generics` to the search path during training, prediction or when printing its `$state`.
 * feat: allow dates in datefeatures pipe op and use data.table for date feature generation.
 * Added support for internal validation tasks to `PipeOpFeatureUnion`.
+* feat: `PipeOpLearnerCV` can reuse the cross-validation models during prediction by averaging their outputs (`resampling.predict_method = "cv_ensemble"`).
 
 # mlr3pipelines 0.9.0
 
@@ -304,4 +305,3 @@
 # mlr3pipelines 0.1.0
 
 * Initial upload to CRAN.
-